# docs/swagger.yaml
basePath: /
definitions:
  domain.Monitor:
    properties:
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
  domain.MonitorStats:
    properties:
      avg_latency_ms:
        type: number
      p95_latency_ms:
        type: number
      total_pings:
        type: integer
      uptime_percentage:
        type: number
    type: object
  handlers.CreateRequest:
    properties:
      url:
        type: string
    type: object
  handlers.CreateResponse:
    properties:
      created_at:
        type: string
      monitor_id:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  handlers.UpdateRequest:
    properties:
      url:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A simple uptime monitoring service API based on Clean Architecture
  title: Uptime Monitor API
  version: "1.0"
paths:
  /healthz:
    get:
      description: Check if the application process is running.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      summary: Liveness Probe
      tags:
      - Operations
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate using an admin password to receive a JWT.
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login to application
      tags:
      - Auth
  /monitors:
    get:
      description: Retrieve a list of all monitors currently being tracked.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Monitor'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List Monitors
      tags:
      - Monitors
    post:
      consumes:
      - application/json
      description: Add a new URL to the uptime monitoring system.
      parameters:
      - description: Monitor Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Monitor
      tags:
      - Monitors
  /monitors/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a monitor using its unique ID.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Monitor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Monitor
      tags:
      - Monitors
    get:
      description: Retrieve details of a specific monitor using its unique ID.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Monitor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Monitor by ID
      tags:
      - Monitors
    put:
      consumes:
      - application/json
      description: Update the URL of an existing monitor in the system.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Monitor Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Monitor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Monitor
      tags:
      - Monitors
  /monitors/{id}/stats:
    get:
      description: Retrieve stats for a monitored URL using its unique ID.
      parameters:
      - description: Monitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MonitorStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fetch Monitor Stats
      tags:
      - Monitors
  /readyz:
    get:
      description: Check if the application is fully ready to accept traffic.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Readiness Probe
      tags:
      - Operations
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by your JWT Token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"