# internal/api/response/response.go
package response

import (
	"encoding/json"
	"net/http"
)

func JSON(w http.ResponseWriter, status int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)

	if data != nil {
		_ = json.NewEncoder(w).Encode(data)
	}
}

type SuccessResponse struct {
	Message string `json:"message"`
}
type ErrorResponse struct {
	Status  int    `json:"status"`
	Message string `json:"message"`
}

func Error(w http.ResponseWriter, status int, message string) {
	JSON(w, status, ErrorResponse{Status: status, Message: message})
}